layout of stories page is changed
